Fixed filter's block visibility
