+++ src/api.ts
@@ -8,8 +8,8 @@ function wait(delay: number) {
 }
 
 export async function getPeople(): Promise<Person[]> {
-  // keep this delay for testing purpose
   return wait(500)
     .then(() => fetch(API_URL))
-    .then(response => response.json());
+    .then(response => response.json())
+    .catch(() => ({ Error: 'Unexpected error' }));
 }
